+++ src/mvvm/awtk/navigator_handler_awtk.h
@@ -48,6 +48,7 @@ navigator_handler_t* navigator_handler_awtk_create(void);
 
 navigator_handler_t* navigator_handler_awtk_toast_create(void);
 navigator_handler_t* navigator_handler_awtk_home_create(void); 
+navigator_handler_t* navigator_handler_awtk_back_create(void); 
 
 #define NAVIGATOR_HANDLER_AWTK(h) ((navigator_handler_awtk_t*)h)