+++ src/mvvm/base/navigator.c
@@ -140,6 +140,10 @@ ret_t navigator_back_to_home(void) {
   return navigator_to(NAVIGATOR_REQ_HOME);
 }
 
+ret_t navigator_back(void) {
+  return navigator_to(NAVIGATOR_REQ_BACK);
+}
+
 ret_t navigator_to_ex(navigator_request_t* req) {
   return navigator_handle_request(navigator(), req);
 }