+++ docs/changes.md
@@ -1,6 +1,7 @@
 # 最新动态
 * 2019/11/02
   * 实现navigator\_toast
+  * 增加navigator\_back
   * 增加navigator\_back\_to\_home
 
 * 2019/10/23